Fix comment links to point to top of comment boxes
Prior to this change, the ID attributes for comment links were set on
the `h2` elements that contain the comment author names. As a result, a
comment link pointed to the author name, not to the top of the comment
box.

This fix moves the ID attributes to the `section` elements that contain
comments, so that clicking a comment link takes us to to the top of the
comment box.