Fix background flicker with dark mode on page load
Prior to this change, the JavaScript code to set the page background to
dark was triggered after the complete page loaded. This caused the page
to load with a white background first and then switch to a dark
background. This caused a very brief but noticeable flicker.

This change fixes this issue by setting the page background to dark as
soon as the `body` element is defined.