Prevent scroll bar appearing due to long word
Here is an example of a page where a scroll bar appears due to a long
word: https://susam.net/comments/fd-100.html

In particular, the comment by MrLeap (currently comment #16 of 38) has
a long link that Safari and Chrome refuse to break into multiple
lines.  This causes the link to overflow the enclosing `<section>`
element while viewing the page on a narrow screen.

This change resolves the problem by ensuring that long words are
always broken into multiple lines, even when there is no acceptable
breaking point within the word.